0x13-more_singly_linked_lists

+++ 0x13-more_singly_linked_lists/task/0-print_listint.c
@@ -8,7 +8,8 @@
 
 size_t print_listint(const listint_t *h)
 {
-    int i = 0;
+    size_t i = 0;
+
     while (h != NULL)
     {
         printf("%i\n", h->n);
